Dash:
-> dash now works when player has no velocity. trouble with gravity
wanting to dash downwards when direction not going in any direction so set
y value to a klooksgood value that also gets the player to jump out of the
water from the surface

Pause:

-> pause menu set to stop game when "Start" button on controller pressed
(or escape on keyboard). Don't have the button links working yet but can
pause and unpause game.